// src/types/News.ts
export interface Article {
  source: {
    id: string;
    name: string;
  };
  author: string;
  title: string;
  description: string;
  url: string;
  urlToImage: string;
  publishedAt: string;
  content: string;
}

export interface INewsList {
  status: string;
  totalResults: number;
  articles: Array<Article>
}